test reorg & TDD initial red / green test
